+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - UPDATE: update to docker-compose 1.15.0
 - UPDATE: move to CircleCI 2.0
 - UPDATE: old Makefile removed
+- UPDATE: licence date update
 
 ## 17.06.0-ce (2017-07-05)
 - UPDATE: update to docker 17.06.0-ce