+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 
-## 17.06.1-ce (2017-08-21)
+## 17.06.1-ce (2017-08-24)
 - UPDATE: update to docker 17.06.1-ce
 - UPDATE: update to docker-compose 1.15.0
 - UPDATE: move to CircleCI 2.0